extract cameracontrol function
